+++ semgrep-core/tests/bash/metavar_equality_var.bash
@@ -1,0 +1,3 @@
+# ERROR:
+out=''
+touch "$out"