+++ semgrep-core/tests/bash/metavar_func_def.bash
@@ -1,0 +1,10 @@
+# ERROR:
+foo() {
+  echo hello
+}
+
+# ERROR:
+function foo() {
+  echo hello
+}
+